backend/requirements.txt
fastapi
uvicorn
pika
python-dotenv
paho-mqtt
